refactor(web): own the group-sharing plan check in a hook
Passing the plan as a boolean let any caller assert an entitlement the
user did not have, and the two callers already derived it differently.
The flag is gone: updateAgentShares now writes what it is handed, and
useUpdateAgentShares decides what to hand it, dropping group shares the
plan does not allow.

The notice no longer says enterprise-only. Group sharing sits behind
Tier.BUSINESS, so it fires for paying Business customers.

AgentShareUpdatePayload moves to types.ts, where the shared types live.
